move classes so they're not spread across garden + misc
