+++ src/main/java/at/hannibal2/skyhanni/features/garden/sensitivity/SensitivityReducer.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.features.garden
+package at.hannibal2.skyhanni.features.garden.sensitivity
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.api.event.HandleEvent
@@ -7,8 +7,7 @@ import at.hannibal2.skyhanni.config.commands.CommandCategory
 import at.hannibal2.skyhanni.config.commands.CommandRegistrationEvent
 import at.hannibal2.skyhanni.config.features.garden.SensitivityReducerConfig
 import at.hannibal2.skyhanni.events.GuiRenderEvent
-import at.hannibal2.skyhanni.features.garden.sensitivity.MouseSensitivityManager
-import at.hannibal2.skyhanni.features.garden.sensitivity.MouseSensitivityManager.SensitivityState
+import at.hannibal2.skyhanni.features.garden.GardenApi
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.ChatUtils
 import at.hannibal2.skyhanni.utils.KeyboardManager.isKeyHeld
@@ -25,9 +24,9 @@ object SensitivityReducer {
 
     private var shouldBeActive = false
     private val isActive get() = isAutoActive || isManualActive
-    private val isAutoActive get() = SensitivityState.AUTO_REDUCED.isActive()
-    private val isManualActive get() = SensitivityState.MANUAL_REDUCED.isActive()
-    private val isMouseLocked get() = SensitivityState.LOCKED.isActive()
+    private val isAutoActive get() = MouseSensitivityManager.SensitivityState.AUTO_REDUCED.isActive()
+    private val isManualActive get() = MouseSensitivityManager.SensitivityState.MANUAL_REDUCED.isActive()
+    private val isMouseLocked get() = MouseSensitivityManager.SensitivityState.LOCKED.isActive()
 
     @HandleEvent
     fun onTick() {
@@ -61,10 +60,10 @@ object SensitivityReducer {
 
         if (!isActive) {
             shouldBeActive = true
-            MouseSensitivityManager.state = SensitivityState.AUTO_REDUCED
+            MouseSensitivityManager.state = MouseSensitivityManager.SensitivityState.AUTO_REDUCED
         } else {
             shouldBeActive = false
-            MouseSensitivityManager.state = SensitivityState.UNCHANGED
+            MouseSensitivityManager.state = MouseSensitivityManager.SensitivityState.UNCHANGED
         }
     }
 
@@ -88,21 +87,21 @@ object SensitivityReducer {
         if (config.onGround.get() && !onGround) return
         if (!isActive) {
             shouldBeActive = true
-            MouseSensitivityManager.state = SensitivityState.AUTO_REDUCED
+            MouseSensitivityManager.state = MouseSensitivityManager.SensitivityState.AUTO_REDUCED
         } else {
             shouldBeActive = false
-            MouseSensitivityManager.state = SensitivityState.UNCHANGED
+            MouseSensitivityManager.state = MouseSensitivityManager.SensitivityState.UNCHANGED
         }
     }
 
     private fun manualToggle() {
         if (!isActive) {
             shouldBeActive = true
-            MouseSensitivityManager.state = SensitivityState.MANUAL_REDUCED
+            MouseSensitivityManager.state = MouseSensitivityManager.SensitivityState.MANUAL_REDUCED
             ChatUtils.chat("§bMouse sensitivity is now lowered. Type /shsensreduce to restore your sensitivity.")
         } else {
             shouldBeActive = false
-            MouseSensitivityManager.state = SensitivityState.UNCHANGED
+            MouseSensitivityManager.state = MouseSensitivityManager.SensitivityState.UNCHANGED
             ChatUtils.chat("§bMouse sensitivity is now restored.")
         }
     }
